maskformer_swin: restructure into demo/, tt/, tests/ subdirectories
Reorganize file structure to match the approved pattern for experimental
models (consistent with swin_s, vgg, dpt_large):

- Move implementation files to tt/ subdirectory
- Move runner.py to demo/ subdirectory
- Update all imports to reflect new paths
- Update README.md runner command paths
