+++ models/experimental/maskformer_swin/demo/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Demo entrypoints for MaskFormer Swin-B."""
