+++ models/experimental/maskformer_swin/tests/test_tt_minimal.py
@@ -7,9 +7,9 @@ import os
 import pytest
 import torch
 
-from models.experimental.maskformer_swin.fallback import MaskFormerFallbackPipeline
-from models.experimental.maskformer_swin.ttnn_compat import ttnn
-from models.experimental.maskformer_swin.weights import (
+from models.experimental.maskformer_swin.tt.fallback import MaskFormerFallbackPipeline
+from models.experimental.maskformer_swin.tt.ttnn_compat import ttnn
+from models.experimental.maskformer_swin.tt.weights import (
     WeightConversionConfig,
     convert_state_dict_to_tt,
     download_reference_weights,
