+++ models/experimental/maskformer_swin/tt/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""TT-NN implementation modules for MaskFormer Swin-B."""
