+++ models/experimental/maskformer_swin/tt/ttnn_compat.py
